# Precompute valid geom collision pairs and implement contact include/exclude