--- a/mujoco_warp/_src/io.py
+++ b/mujoco_warp/_src/io.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 # ==============================================================================
 
-from typing import Optional
-from typing import Tuple
+from typing import Optional, Tuple
 
 import mujoco
 import numpy as np