this pr precomputes valid geom collision pairs with the function `collision_driver.geom_pair`. additional filtering for contact exclude is included with this function. the resulting list of geom pairs is computed in `put_model` and is available as the field `nxn_geom_pair` in `Model`.

benchmarking:
```
mjwarp-testspeed --function=collision --mjcf=collision.xml --batch_size=8192
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.01 s
 Total simulation time: 0.12 s
 Total steps per second: 67,267,046
 Total realtime factor: 134,534.09 x
 Total time per step: 14.87 ns
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.01 s
 Total simulation time: 0.12 s
 Total steps per second: 65,986,259
 Total realtime factor: 131,972.52 x
 Total time per step: 15.15 ns
```