@erikfrey 2025-03-21
Hmm!  This is an interesting design quesiton, I think there's a few pros and cons.  I'll list them out here later today and then we can discuss.