@erikfrey 2025-03-28
OK!  Very interesting!

Pros of precomputing valid geom pairs:
* We can skip ever even considering invalid geom pairs, e.g. scenes with many visual geoms will be a bit faster because broadphase never even visits that pair
* `<pair>` include and exclude is a bit easier to insert/delete statically than insert/delete on the fly.  MuJoCo efficiently merges broadphase results and pair contacts by ensuring both are sorted, then it can do a linear merge, in order to ensure we don't have duplicate contacts.  If we do this dynamically we'll have to futz with it a bit to ensure it's fast.

Cons:
* We're reifying an array of geom pairs that is ngeom^2.  Pragmatically this is probably not a big deal, even at 10,000 geoms that's only like 500MB or so on device.  So this is not an enormous con.
* If user changes contype or conaffinity, we must now re-JIT the physics step, whereas the existing version has no such requirement.

I think I could go either way.  Curious what you all think.