@thowell 2025-04-03
i think the identified cons are minor compared to the pros for the precomputed geom pairs approach.

- based on the estimate of ~500MB for 10,000 geoms, the memory requirements for the precomputed geom pairs seem reasonable
- if the model is changed (eg, contype/conaffinity modified) i think it is reasonable to expect that the physics step should be recompiled