@adenzler-nvidia 2025-04-03
I would like to do a deep-dive and see how the performance profile changes with this, but that's more for my own education..

I'm with Taylor on this one. Given that we have the (somewhat arbitrary) threshold that switches to SAP broadphase for large geom counts, I think it's ok to have this list as it will not be very large in many cases.