@erikfrey 2025-04-04
OK!  Then I'll list two more design requests before starting to look carefully at the code:

1) Can we also add support for contact/pair in this PR?  This will help us fully address #115 and seems related enough, but let me know if you feel otherwise @thowell 

2) Can we reuse this array in the sweep and prune broadphase?  Pair and exclude are just tricky enough that we ought to reuse the same logic in both approaches if possible.