@thowell 2025-04-06
@erikfrey yes will add support for contact pair to this pr