@erikfrey 2025-04-08
@btaba just for simplicity - you're right that in C MuJoCo, `mj_broadphase` does SAP at the body level and then compares bounding spheres for midphase - it's not clear that following these steps will be performant for GPU.  We already have a SAP broadphase but my understanding is sorting is currently a bottleneck, in practice we've found that n^2 bounding sphere test is performant for at least `ngeom` <= 100 - haven't tested higher yet.