@btaba 2025-04-08
@erikfrey Thanks for context; I wasn't necessarily mentioning SAP in particular, just mentioning broadphase on bodies, midphase on geoms, rather than broadphase on geoms, no midphase. The full version I think would involve a CPU callback [something like this](https://github.com/btaba/mujoco/blob/06f5d7c0b0cfed27007cf19a2d45736edfd2e212/mjx/mujoco/mjx/_src/cuda/engine_collision_driver.cu.cc#L513-L514), but it might still be performant esp. if different parts of the mj_step graph can run in parallel (multiple streams), and it may scale better in the number of bodies/geoms