@adenzler-nvidia 2025-04-09
It's always hard to have an intuition about performance, but my intuition is that you need a lot of potential geom pairs to justify the need for a midphase. Most of the time it's better to expose as much parallelism as possible, especially looking at potential future bigger GPUs. But of course only measuring this will tell us where the tradeoff is.

There are still optimizations on the table with n2 because in theory you only need to load echo geom once and can cache it in shared memory - that likely gives another nice boost.

Running with multiple streams is cool as well but I would only go there once we have a more or less complete port ready. These things are super hard to debug, and if you have good GPU utilization it likely isn't really helping a lot with performance.