@thowell 2025-04-09
@erikfrey support for contact pairs has been added