@thowell 2025-04-15
@erikfrey `io.geom_pair` now generates `ngeom * (ngeom - 1) / 2` pair ids and geom pairs:
- `id == -2`: pair should be skipped during broadphase
- `id == -1`: pair is evaluated during broadphase and contact parameters are mixed
- `id > -1`: pair is evaluated during broadphase and contact parameters are specified via contact pair

`nxn_broadphase` and `sap_broadphase` both utilize `nxn_pairid` to filter geom pairs